balanced interval Tree

+++ src/models/Alignment.java
@@ -5,7 +5,7 @@ import utils.DatabaseConnector;
 
 import java.sql.SQLException;
 
-public record Alignment(String sseqid, int qstart, int qend, float eval, int bitscore) {
+public record Alignment(String sseqid, int qstart, int qend, float eval, float bitscore) {
     @Override
     public String sseqid() {
         try {
